requirements.txt
ipdb
python-dotenv
requests
pytest